feat(chat): extract code blocks from MarkdownText into CodeBlockCard (#664)
- Add language field to MdBlock.Code data class
- Parse language from ```fence markers in parseBlocks()
- Replace old private CodeBlockCard with new one from MessageCards.kt
  (syntax highlighting + language badge + copy checkmark)
- Remove ~60 lines of duplicated code block rendering from MarkdownText
